requirements.txt
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.20
celery==3.1.18
celery-with-redis==3.0
Django==1.8.4
django-celery==3.1.16
django-filter==0.11.0
django-model-utils==2.3.1
django-pandas==0.3.0
djangorestframework==3.2.2
geopy==1.11.0
jsonfield==1.0.3
kombu==3.0.26
lxml==3.4.4
Markdown==2.6.2
numpy==1.9.2
pandas==0.16.2
python-dateutil==2.4.2
pytz==2015.4
redis==2.10.3
requests==2.7.0
six==1.9.0
